first solution using external TSP

+++ src/main/java/com/pld/agile/model/tsp/Graph.java
@@ -11,7 +11,7 @@ public interface Graph {
 	 * @param j 
 	 * @return the cost of arc (i,j) if (i,j) is an arc; -1 otherwise
 	 */
-	public abstract int getCost(int i, int j);
+	public abstract Double getCost(int i, int j);
 
 	/**
 	 * @param i
